jingra/pyproject.toml
[project]
name = "jingra"
version = "1.0.0"
description = "Comprehensive benchmarking tool for various search, vector search, observability, and security use cases across multiple engines."
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Benchmark Team"}
]
keywords = ["benchmark", "search", "vector-search", "observability", "security", "elasticsearch", "opensearch", "retrieval"]

dependencies = [
    # Search engine clients
    "elasticsearch>=8.0.0",
    "opensearch-py>=2.0.0",
    "qdrant-client>=1.9.0",
    # Data handling
    "datasets>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    # Configuration
    "pyyaml>=6.0",
    # HTTP
    "requests>=2.28.0",
    # Progress bars
    "tqdm>=4.65.0",
    # Plotting
    "matplotlib>=3.10.8",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "jingra[dev]",
]

[project.scripts]
jingra = "src.benchmark.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/benchmark"]
omit = [
    "src/benchmark/plotting/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "@abstractmethod",
]
